test: move colocated specs into their directories' __tests__
`check:test-placement` landed in 46d92df6f and fails on develop: four specs
sit beside their subjects in directories that had already settled on
`__tests__/`, so three directories mix both conventions and every pull request
red-lights on the gate.

The gate commit was authored on 2026-08-23 and merged on 2026-08-27 (#886).
In the four days it waited, #897, #952 and #960 each added a colocated spec —
correctly, since no gate existed on develop yet. Nothing conflicted textually,
so the merge was clean and the rule was violated the moment it arrived.

Move the four files and re-root their relative specifiers one level. No test
body changes, no subject changes: same 15 assertions, same modules under test.

Directories fixed:
  src/components/ComposerInput            (10 already in __tests__/)
  src/modules/WorkStation/shared/StatusBar (1 already in __tests__/)
  src/util/language                        (2 already in __tests__/)
